requirements.txt
Django==2.2.4
django-compat==1.0.15
django-filter==2.2.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
Markdown==3.1.1
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
